+++ packages/components/pages/f-checkout/CHANGELOG.md
@@ -3,6 +3,17 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v3.2.0
+------------------------------
+*November 9, 2021*
+
+### Added
+`CheckoutFormField` component.
+
+### Changed
+- Guest Component to use `CheckoutFormField` component.
+
+
 v3.1.1
 ------------------------------
 *November 5, 2021*